moved EMPTY_MERKLE_TREE to raiden.transfer.state
