+++ raiden/tests/smart_contracts/netting_channel/test_close.py
@@ -4,16 +4,14 @@ from ethereum.tools import tester
 from ethereum.tools.tester import TransactionFailed
 from coincurve import PrivateKey
 
-from raiden.messages import (
-    EMPTY_MERKLE_ROOT,
-    DirectTransfer,
-)
+from raiden.messages import DirectTransfer
 from raiden.tests.utils.messages import (
     make_direct_transfer,
 )
 from raiden.tests.utils.transfer import make_direct_transfer_from_channel
 from raiden.utils import privatekey_to_address, sha3, event_decoder, address_encoder
 from raiden.tests.utils.factories import make_address
+from raiden.transfer.state import EMPTY_MERKLE_ROOT
 
 
 def test_close_event(tester_chain, tester_nettingcontracts, tester_events):
